+++ lib/src/overlays/qoverlay.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/cupertino.dart';
+
+// ignore: one_member_abstracts
+abstract class QOverlay {
+  Future<T?> show<T>({NavigatorState? state, BuildContext? context});
+}