+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix notFoundPage path in nested navigator
 - Add `AddRoutes` and `RemoveRoute`
 - Route.path can be without slash at the start.
+- Add `QDialog`, so user can open dialogs.
 
 ## [1.1.4] 12.04.2021